requirements.txt
requests
python-dotenv
python-docx
pypdf2
pandas
openpyxl
flask
